feat: implement uf_completed
